requirements.txt
flask>=3.0
openai>=1.14
python-dotenv>=1.0

